UPDATE : user.service.js (sans oublier le git add cette fois)

+++ back/src/service/user.service.js
@@ -1,33 +1,68 @@
 // Imports
+const mongoose = require("mongoose") ;
 const UserRepository = require("../repository/user.repository") ;
 
-
+function makeError(message, statusCode = 400) {
+    const err = new Error(message) ;
+    err.statusCode = statusCode ;
+    return err ;
+}
     
 function getAllUsers() {
     return UserRepository.findAllUsers() ;
 }
 
 function getUserById(id) {
-    return UserRepository.findUserById(id) ;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw makeError("ID invalide", 400) ;
+    }
+
+    const user = UserRepository.findUserById(id) ;
+    if (!user) {
+        throw new makeError("User not found", 404) ;
+    }
+    return user ;
 }
 
 function getUserByName(name) {
-    return UserRepository.findUserByName(name) ;
+    const user = UserRepository.findUserByName(name) ;
+    if (!user) {
+        throw new makeError("User not found", 404) ;
+    }
+    return user ;
 }
 
 function getUserByEmail(email) {
-    return UserRepository.findUserByEmail(email) ;
+    const user = UserRepository.findUserByEmail(email) ;
+    if (!user) {
+        throw new makeError("User not found", 404) ;
+    }
+    return user ;
 }
 
 function createUser(userData) {
-    return UserRepository.createUser(userData) ;
+    const user = UserRepository.createUser(userData) ;
+    if (!user) {
+        throw new makeError("Failed to create user", 500) ;
+    }
+    return user ;
 }
 
 function updateUser(id, userData) {
-    return UserRepository.updateUser(id, userData) ;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw makeError("ID invalide", 400) ;
+    }
+    const user = UserRepository.updateUser(id, userData) ;
+    if (!user) {
+        throw new makeError("Failed to update user", 500) ;
+    }
+    return user ;
 }
 
 function deleteUser(id) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        throw makeError("ID invalide", 400) ;
+    }
     return UserRepository.deleteUser(id) ;
 }
